+++ changelog.md
@@ -7,6 +7,7 @@
 * Fixed villager marriage limit math
 * You can no longer restrict bells or gravestones, causing them to get stuck
 * Vanilla player models no longer modify your eye height
+* Player eye height now automatically refreshes on world join and editor changes
 
 # 7.5.9